@@ -13,16 +13,17 @@
         "@sentry/tracing": "^6.16.1",
         "@sentry/vue": "^6.16.1",
         "@symfony/webpack-encore": "^1.7.0",
-        "@vue/composition-api": "^1.4.3",
+        "@vue/compat": "^3.1.5",
+        "@vue/compiler-sfc": "^3.1.5",
         "axios": "^0.24.0",
         "axios-cache-adapter": "^2.7.3",
-        "bootstrap": "^4.6.1",
-        "bootstrap-vue": "^2.21.2",
-        "chart.js": "^2.9.4",
-        "core-js": "^3.20.2",
+        "bootstrap": "^5.1.3",
+        "bootstrap-vue-3": "^0.1.0",
+        "chart.js": "^3.7.0",
+        "core-js": "^3.20.3",
         "eslint": "^7.32.0",
         "eslint-loader": "^4.0.2",
-        "eslint-plugin-vue": "^7.20.0",
+        "eslint-plugin-vue": "^8.3.0",
         "file-loader": "^6.2.0",
         "js-yaml": "^4.1.0",
         "localforage": "^1.10.0",
@@ -31,18 +32,17 @@
         "pinia": "^2.0.9",
         "popper.js": "^1.16.0",
         "regenerator-runtime": "^0.13.9",
-        "sass": "^1.45.2",
+        "sass": "^1.48.0",
         "sass-loader": "^12.4.0",
         "timeago.js": "^4.0.2",
-        "vue": "^2.6.14",
-        "vue-chartjs": "^3.5.1",
+        "vue": "^3.1.5",
         "vue-context": "^6.0.0",
-        "vue-i18n": "^8.26.8",
+        "vue-i18n": "^9.1.9",
         "vue-i18n-extract": "^2.0.0",
-        "vue-loader": "^15.9.8",
+        "vue-loader": "^16.8.3",
         "vue-mapbox": "^0.4.1",
         "vue-slicksort": "^1.1.3",
-        "vue-template-compiler": "^2.6.14",
+        "vue-star-rating": "^2.1.0",
         "webpack-bundle-analyzer": "^4.5.0",
         "webpack-notifier": "^1.15.0"
     },
@@ -53,5 +53,8 @@
         "dev": "encore dev",
         "watch": "NODE_ENV=development encore dev-server --hot",
         "build": "encore production --progress"
+    },
+    "dependencies": {
+        "vue-chart-3": "^3.0.7"
     }
 }
