@@ -1,6 +1,7 @@
 <template>
   <component
     :is="component"
+    ref="root"
     v-bind="props"
   />
 </template>
@@ -11,8 +12,10 @@ import BookstoresAdmin from "./admin/BookstoresAdmin";
 import EdgeProgress from "./admin/EdgeProgress";
 import Site from "./Site";
 import Privacy from "./Privacy";
-import { mapActions } from "pinia";
-import { l10n } from "../stores/l10n";
+
+const {l10n: l10nStore} = require("../stores/l10n");
+
+import {ref, onMounted} from 'vue'
 
 export default {
   name: "App",
@@ -23,19 +26,39 @@ export default {
     Privacy,
     Site
   },
-  data() {
-    const {component, props} = this.$attrs
+  setup() {
+    const
+      component = ref(null),
+      props = ref({}),
+      root = ref(null),
+      {loadL10n} = l10nStore()
+    onMounted(async () => {
+      for (const {name, value} of root.value.parentElement.attributes) {
+        if (name === 'component') {
+          component.value = value
+        } else {
+          props.value[name] = value
+        }
+      }
+      await loadL10n()
+    })
+
     return {
+      root,
       component,
-      props,
+      props
     }
   },
-  async mounted() {
-    await this.loadL10n()
-  },
-  methods: {
-    ...mapActions(l10n, ['loadL10n'])
-  }
+  // data() {
+  //   const {component, props} = this.$attrs
+  //   return {
+  //     component,
+  //     props,
+  //   }
+  // },
+  // async mounted() {
+  //   await this.loadL10n()
+  // },
 }
 </script>
 
