@@ -1,14 +1,25 @@
+<template>
+  <BarChart
+    v-if="chartData"
+    :chart-data="chartData"
+    :options="options"
+  />
+</template>
+
 <script>
-import {HorizontalBar} from 'vue-chartjs'
 import collectionMixin from "../../mixins/collectionMixin";
 import {mapActions, mapState} from "pinia";
 import l10nMixin from "../../mixins/l10nMixin";
 import { collection } from "../../stores/collection";
 import { coa } from "../../stores/coa";
+import {BarChart} from "vue-chart-3";
+
+import {ArcElement, Chart, CategoryScale, LinearScale, Legend, BarElement, BarController, Title, Tooltip} from 'chart.js';
+Chart.register(Legend, CategoryScale, BarElement, LinearScale, BarController, Tooltip, Title, ArcElement);
 
 export default {
   name: "PossessionStats",
-  extends: HorizontalBar,
+  components: {BarChart},
   mixins: [collectionMixin, l10nMixin],
 
   props: {
@@ -20,7 +31,8 @@ export default {
   emits: ['change-dimension'],
 
   data: () => ({
-    orientation: 'vertical'
+    chartData: null,
+    options: {}
   }),
 
   computed: {
@@ -65,14 +77,14 @@ export default {
     labels: {
       immediate: true,
       async handler(newValue) {
-        this.$emit('change-dimension', this.orientation === 'vertical' ? 'height' : 'width', 100 + 30 * newValue.length)
-        this.$emit('change-dimension', this.orientation === 'vertical' ? 'width' : 'height', 500)
+        this.$emit('change-dimension', 'height', 100 + 30 * newValue.length)
+        this.$emit('change-dimension', 'width', 500)
       }
     },
     values: function (newValue) {
       if (newValue) {
         const vm = this
-        this.renderChart({
+        this.chartData = {
           datasets: [
             {
               data: this.values[0],
@@ -89,43 +101,50 @@ export default {
           ],
           labels: this.labels,
           legends: [this.$t("Numéros possédés"), this.$t("Numéros référencés non-possédés")]
-        }, {
-          title: {
-            display: true,
-            text: this.$t("Possession des numéros")
-          },
+        }
+
+        this.options = {
           responsive: true,
+          indexAxis: 'y',
           maintainAspectRatio: false,
           scales: {
-            xAxes: [{
+            x: {
+              min: 0,
+              max: vm.unit === 'percentage' ? 100 : undefined,
               stacked: true,
               ticks: {
                 stepSize: 1,
                 callback: value => vm.unit === 'percentage' ? `${value}%` : value
               }
-            }],
-            yAxes: [{
+            },
+            y: {
               stacked: true
-            }]
+            }
           },
-          tooltips: {
-            enabled: true,
-            position: 'nearest',
-            mode: 'index',
-            axis: this.orientation === 'vertical' ? 'y' : 'x',
-            intersect: false,
-            callbacks: {
-              title: ([tooltipItem]) => {
-                const publicationcode = tooltipItem[vm.orientation === 'vertical' ? 'yLabel' : 'xLabel'];
-                if (!vm.publicationNames[publicationcode]) {
-                  vm.publicationNames[publicationcode] = '?';
-                }
-                return `${vm.publicationNames[publicationcode] || '?'} (${vm.countryNames[publicationcode.split('/')[0]]})`;
-              },
-              label: (tooltipItems, {datasets}) => `${datasets[tooltipItems.datasetIndex].legend} : ${tooltipItems[vm.orientation === 'vertical' ? 'xLabel' : 'yLabel']}${vm.unit === 'percentage' ? '%' : ''}`
+          plugins: {
+            title: {
+              display: true,
+              text: this.$t("Possession des numéros")
+            },
+            tooltip: {
+              enabled: true,
+              position: 'nearest',
+              mode: 'index',
+              axis: 'y',
+              intersect: false,
+              callbacks: {
+                title: ([tooltipItem]) => {
+                  const publicationcode = tooltipItem.label
+                  if (!vm.publicationNames[publicationcode]) {
+                    vm.publicationNames[publicationcode] = '?';
+                  }
+                  return `${vm.publicationNames[publicationcode] || '?'} (${vm.countryNames[publicationcode.split('/')[0]]})`;
+                },
+                label: (tooltipItem) => `${tooltipItem.dataset.label}: ${tooltipItem.raw}${vm.unit === 'percentage' ? '%' : ''}`,
+              }
             }
           }
-        })
+        }
       }
     }
   },
